Refactor out  parsing
Needed to make source field in Schedule public. Need to specify it in Schedule::new, but that means carrying the string through the parser.
