+++ src/lib.rs
@@ -45,6 +45,7 @@ mod time_unit;
 mod ordinal;
 mod specifier;
 mod queries;
+mod parsing;
 
 pub use crate::schedule::Schedule;
 pub use crate::time_unit::TimeUnitSpec;
